layers: Update SPIRV-Tools
